Refactor import statements in buyer, party, type, marketing, section, properties, section, lab_dip, factory, merchandiser, pi_cash, recipe, info, stock_to_sfg, challan, machine queries to include decimalToNumber function from variables.js

+++ src/db/commercial/query/pi_cash.js
@@ -9,6 +9,7 @@ import {
 import * as hrSchema from '../../hr/schema.js';
 import db from '../../index.js';
 import * as publicSchema from '../../public/schema.js';
+import { decimalToNumber } from '../../variables.js';
 
 import { alias } from 'drizzle-orm/pg-core';
 import { bank, lc, pi_cash, pi_cash_entry } from '../schema.js';
@@ -176,9 +177,9 @@ export async function selectAll(req, res, next) {
 			pi_cash.updated_at,
 			pi_cash.remarks,
 			pi_cash.is_pi,
-			pi_cash.conversion_rate,
-			pi_cash.weight,
-			pi_cash.receive_amount
+			pi_cash.conversion_rate::float8,
+			pi_cash.weight::float8,
+			pi_cash.receive_amount::float8
 		FROM 
 			commercial.pi_cash
 		LEFT JOIN
